requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-microsoft-azure==8.2.0rc1
apache-airflow-providers-microsoft-mssql